requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
oauthlib==3.2.0
pipreqs==0.4.11
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
yarg==0.1.9
